Moved linear and conv layer under 'nn' folder - similar to PyTorch
